pyproject.toml
[project]
name = "eplumber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "appdirs",
    "typing-extensions",
    "pydantic",
    "paho-mqtt>=2.1.0",
    "requests>=2.32.3",
    "jsonpath-ng",
    "ipdb>=0.13.13",
]
